detect: support projects with non-default build file names
With Gradle, it's possible to have a valid project where no file named `build.gradle` is used.
For these projects, you set the `buildFileName` on the `ProjectDescriptor` in your `settings.gradle`.
The Ratpack project is currently an example of this type of project.
Thus, I've changed the detect logic to consider an application as a Gradle application when either
`build.gradle` or `settings.gradle` is found.

http://www.gradle.org/docs/current/dsl/org.gradle.api.initialization.Settings.html
http://www.gradle.org/docs/current/javadoc/org/gradle/api/initialization/ProjectDescriptor.html
https://github.com/ratpack/ratpack